feat: Rol enum, LocalDateTime, token format kontrolü, header required düzenlemesi

+++ src/main/java/com/cerensahin/sosyalmedya/dto/GonderiDetay.java
@@ -10,9 +10,9 @@ public class GonderiDetay {
     private String icerik;
     private String medyaUrl;
     private String medyaTipi;
-    private String medyaBase64; // YENİ
+    private String medyaBase64;
     private LocalDateTime olusturmaZamani;
-    private int goruntulenmeSayisi;
+    private long goruntulenmeSayisi;
     private long begeniSayisi;
     private long yorumSayisi;
     private List<YorumGorunum> yorumlar;
@@ -27,11 +27,12 @@ public class GonderiDetay {
         this.medyaTipi = g.getMedyaTipi().name();
         this.medyaBase64 = g.getMedyaBase64();
         this.olusturmaZamani = g.getOlusturmaZamani();
-        this.goruntulenmeSayisi = g.getGoruntulenmeSayisi();
+        this.goruntulenmeSayisi = g.getGoruntulenmeSayisi(); // ✅ artık aynı tür
         this.begeniSayisi = begeni;
         this.yorumSayisi = yorumSayisi;
         this.yorumlar = yorumlar;
     }
+
 
     public Long getId() { return id; }
     public Long getKullaniciId() { return kullaniciId; }
@@ -40,7 +41,7 @@ public class GonderiDetay {
     public String getMedyaTipi() { return medyaTipi; }
     public String getMedyaBase64() { return medyaBase64; }
     public LocalDateTime getOlusturmaZamani() { return olusturmaZamani; }
-    public int getGoruntulenmeSayisi() { return goruntulenmeSayisi; }
+    public long getGoruntulenmeSayisi() { return goruntulenmeSayisi; }
     public long getBegeniSayisi() { return begeniSayisi; }
     public long getYorumSayisi() { return yorumSayisi; }
     public List<YorumGorunum> getYorumlar() { return yorumlar; }
